// app/types/Establishment.ts
export interface Establishment {
    id?: string;
    logo?: string;
    cnpj: string;
    nome: string;
    telefone: string;
    site?: string;
    email: string;
    emailFinanceiro?: string;
    cep?: string;
    endereco?: string;
    numero?: string;
    bairro?: string;
    complemento?: string;
    cidade?: string;
    estado?: string;
    status?: string;
  }